adding salmon

+++ R/salmonIndex.R
@@ -1,19 +1,24 @@
-#' @title A skeleton function to handle a docker containier
-#' @description This function executes a ubuntu docker that produces as output helloworld.txt file
+#' @title A function to create a Salmon pseudo reference
+#' @description This function executes the Salmon docker that produces as output a transcripts index file.
 #' @param group, a character string. Two options: sudo or docker, depending to which group the user belongs
-#' @param scratch.folder, a character string indicating the path of the scratch folder
-#' @param data.folder, a character string indicating the folder where input data are located and where output will be written
-#' @author Name Family name, myemail [at] somewhere [dot] org, Affiliation
+#' @param index.folder, a character string indicating the folder where transcriptime index will be created.
+#' @param ensembl.ensembl.urltranscriptome, a character string indicating the URL from ENSEMBL ftp for the transcripts fasta file of interest
+#' @param ensembl.urlgtf, a character string indicating the URL from ENSEMBL ftp for the GTF for genome of interest
+#' @param k, a number indicating the k-mers length, 31 eems to work well for reads of 75bp or longer, but you might consider a smaller k if dealing with shorter reads.
+#' @author Raffaele Calogero, raffaele.calogero [at] unito [dot] it, Bioinformatics and Genomics unit University of Torino Italy
 #' 
 #' @examples
+#' index -t transcripts.fa -i transcripts_index --type quasi -k 31
 #' \dontrun{
 #'     #running skeleton
-#'     skeleton(group="docker", scratch.folder="/Users/raffaelecalogero/Desktop/scratch", 
-#'     data.folder=getwd())
+#'     salmonIndex(group="docker", index.folder=getwd(), 
+#'     ensembl.urltranscriptome="ftp://ftp.ensembl.org/pub/release-90/fasta/mus_musculus/cdna/Mus_musculus.GRCm38.cdna.all.fa.gz",
+#'     ensembl.urlgtf="ftp://ftp.ensembl.org/pub/release-90/gtf/mus_musculus/Mus_musculus.GRCm38.90.gtf.gz", 
+#'     k=31)
 #' }
 #'
 #' @export
-skeleton <- function(group=c("sudo","docker"), scratch.folder, data.folder){
+salmonIndex <- function(group=c("sudo","docker"), index.folder, ensembl.urltranscriptome, ensembl.urlgtf, k=31){
 
   #testing if docker is running
   test <- dockerTest()
@@ -26,36 +31,28 @@ skeleton <- function(group=c("sudo","docker"), scratch.folder, data.folder){
   #running time 1
   ptm <- proc.time()
   #setting the data.folder as working folder
-  if (!file.exists(data.folder)){
-    cat(paste("\nIt seems that the ",data.folder, " folder does not exist\n"))
+  if (!file.exists(index.folder)){
+    cat(paste("\nIt seems that the ",index.folder, " folder does not exist\n"))
     return(2)
   }
-  setwd(data.folder)
-  #check  if scratch folder exist
-  if (!file.exists(scratch.folder)){
-    cat(paste("\nIt seems that the ",scratch.folder, " folder does not exist\n"))
-    return(3)
-  }
-  tmp.folder <- gsub(":","-",gsub(" ","-",date()))
-  scrat_tmp.folder=file.path(scratch.folder, tmp.folder)
-  writeLines(scrat_tmp.folder,paste(data.folder,"/tempFolderID", sep=""))
-  cat("\ncreating a folder in scratch folder\n")
-  dir.create(file.path(scrat_tmp.folder))
+  setwd(index.folder)
+  system(paste("wget -O transcripts.fa.gz ", ensembl.urltranscriptome, sep=""))
+  system("gzip -d transcripts.fa.gz")
+  system(paste("wget -O genome.gtf.gz ",ensembl.urlgtf, sep=""))
+  system("gzip -d genome.gtf.gz")
   #executing the docker job
   if(group=="sudo"){
-    params <- paste("--cidfile ",data.folder,"/dockerID -v ",scrat_tmp.folder,":/scratch -v ", data.folder, ":/data -d docker.io/repbioinfo/ubuntu sh /bin/skeleton.sh", sep="")
-    resultRun <- runDocker(group="sudo",container="repbioinfo/salmon.2017.01", params=params)
+#    params <- paste("--cidfile ",index.folder,"/dockerID -v ", index.folder, ":/index -d docker.io/repbioinfo/salmon.2017.01 /usr/local/bin/salmon index -t /index/transcripts.fa -i /index/transcripts_index --type quasi -k ",k, sep="")
+    resultRun <- runDocker(group="sudo",container="docker.io/repbioinfo/salmon.2017.01", params=params)
   }else{
-    params <- paste("--cidfile ",data.folder,"/dockerID -v ",scrat_tmp.folder,":/scratch -v ", data.folder, ":/data -d docker.io/repbioinfo/ubuntu sh /bin/skeleton.sh", sep="")
-    resultRun <- runDocker(group="docker",container="repbioinfo/salmon.2017.01", params=params)
+#    params <- paste("--cidfile ",index.folder,"/dockerID -v ", index.folder, ":/index -d docker.io/repbioinfo/salmon.2017.01 /usr/local/bin/salmon index -t /index/transcripts.fa -i /index/transcripts_index --type quasi -k ",k, sep="")
+     params <- paste("--cidfile ",index.folder,"/dockerID -v ", index.folder, ":/index -d docker.io/repbioinfo/salmon.2017.01 sh /bin/salmon_index.sh ",k, sep="")
+     resultRun <- runDocker(group="docker",container="docker.io/repbioinfo/salmon.2017.01", params=params)
   }
   #waiting for the end of the container work
-  if(!resultRun){
-    system(paste("cp ", scrat_tmp.folder, "/* ", data.folder, sep=""))
-  }
   #running time 2
   ptm <- proc.time() - ptm
-  dir <- dir(data.folder)
+  dir <- dir(index.folder)
   dir <- dir[grep("run.info",dir)]
   if(length(dir)>0){
     con <- file("run.info", "r")
@@ -75,15 +72,17 @@ skeleton <- function(group=c("sudo","docker"), scratch.folder, data.folder){
   }
 
   #saving log and removing docker container
-  container.id <- readLines(paste(data.folder,"/dockerID", sep=""), warn = FALSE)
-  system(paste("docker logs ", substr(container.id,1,12), " &> ",data.folder,"/", substr(container.id,1,12),".log", sep=""))
+  container.id <- readLines(paste(index.folder,"/dockerID", sep=""), warn = FALSE)
+  system(paste("docker logs ", substr(container.id,1,12), " &> ",index.folder,"/", substr(container.id,1,12),".log", sep=""))
   system(paste("docker rm ", container.id, sep=""))
   #removing temporary folder
   cat("\n\nRemoving the temporary file ....\n")
-  system(paste("rm -R ",scrat_tmp.folder))
-  system("rm -fR out.info")
-  system("rm -fR dockerID")
-  system("rm  -fR tempFolderID")
-  system(paste("cp ",paste(path.package(package="docker4seq"),"containers/containers.txt",sep="/")," ",data.folder, sep=""))
+  if(resultRun=="false"){
+     system("rm -fR dockerID")
+     system("rm  -fR tempFolderID")
+     system(paste("cp ",paste(path.package(package="docker4seq"),"containers/containers.txt",sep="/")," ",data.folder, sep=""))
+  }else{
+    cat(paste("there was an error in the execution of ", substr(container.id,1,12), "\n"), sep="")
+  }
   setwd(home)
 }
